refactor: post comments
